+++ classes/baseDecorator.rb
@@ -1,0 +1,14 @@
+require './nameable'
+
+# Base Decorator Class
+
+class BaseDecorator < Nameable
+  def initialize(nameable)
+    super()
+    @nameable = nameable
+  end
+
+  def correct_name
+    @nameable.correct_name
+  end
+end
