+++ classes/capitalizeDecorator.rb
@@ -1,0 +1,7 @@
+# Prepare CapitalizeDecorator
+
+class CapitalizeDecorator < BaseDecorator
+  def correct_name
+    @nameable.correct_name.capitalize
+  end
+end
